Refactoring the unittest file
